adding theme docs and refactoring stories
