+++ src/js/components/CheckBox/stories/Basics.js
@@ -1,0 +1,36 @@
+import React, { Component } from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Box, Grommet, CheckBox } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+class SimpleCheckBox extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { checked: !!props.checked };
+  }
+
+  onChange = event => this.setState({ checked: event.target.checked });
+
+  render() {
+    const { checked } = this.state;
+    return (
+      <Grommet theme={grommet}>
+        <Box align="center" pad="large">
+          <CheckBox
+            {...this.props}
+            checked={checked}
+            onChange={this.onChange}
+          />
+        </Box>
+      </Grommet>
+    );
+  }
+}
+
+storiesOf('CheckBox', module)
+  .add('Simple', () => <SimpleCheckBox label="Choice" />)
+  .add('Toggle', () => <SimpleCheckBox label="Choice" toggle />)
+  .add('Disabled', () => <SimpleCheckBox label="Choice" checked disabled />)
+  .add('Reverse', () => <SimpleCheckBox label="Choice" reverse />)
+  .add('No Label', () => <SimpleCheckBox />);
